+++ iCon/Classes/ViewModel/Helpers/RemoteCluster.cs
@@ -1,17 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel;
-using System.Diagnostics.Metrics;
 using System.IO;
-using System.Linq;
 using System.Net.Sockets;
-using System.Reflection;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows;
-using System.Windows.Controls;
 using System.Windows.Threading;
-using Org.BouncyCastle.Asn1.Mozilla;
 using Renci.SshNet;
 using Renci.SshNet.Common;